fix ambigous std::abs overloads
