+++ dune/xt/common/fix-ambiguous-std-math-overloads.cc
@@ -5,23 +5,25 @@
 //      or  GPL-2.0+ (http://opensource.org/licenses/gpl-license)
 //          with "runtime exception" (http://www.dune-project.org/license.html)
 // Authors:
-//   Felix Schindler (2013 - 2014, 2016 - 2017)
-//   René Fritze     (2013, 2015 - 2016, 2018)
-//   Tobias Leibner  (2018)
+//   Felix Schindler (2018)
 
 #include "config.h"
 
-#include "math.hh"
+#include "fix-ambiguous-std-math-overloads.hh"
 
-namespace Dune {
-namespace XT {
-namespace Common {
+namespace std {
 
 
-// this is just for the test to compile
-const std::string Epsilon<std::string, false>::value = "_";
+long unsigned int abs(const long unsigned int& value)
+{
+  return value;
+}
 
 
-} // namespace Common
-} // namespace XT
-} // namespace Dune
+unsigned char abs(unsigned char value)
+{
+  return value;
+}
+
+
+} // namespace std
